#CLK32-1038 : -  logic create message logic
