+++ Packages/ChatSecure/ChatSecure/Classes/Constants.swift
@@ -10,6 +10,8 @@ import Foundation
 enum Constants {
 	static let keySaveTurnServerUser = "keySaveTurnServerUser"
 	static let keySaveTurnServerPWD = "keySaveTurnServerPWD"
+	static let keySaveTurnServer = "keySaveTurnServer"
+	static let keySaveStunServer = "keySaveStunServer"
 	
 	static let senderDeviceId = 111
 	static let receiverDeviceId = 222
