+++ Packages/ChatSecure/ChatSecure/Classes/WebRTCJanus/JanusRole.swift
@@ -105,17 +105,19 @@ public class JanusRole: JanusPlugin {
 		
 		let turnUser = UserDefaults.standard.string(forKey: ChatSecure.Constants.keySaveTurnServerUser) ?? ""
 		let turnPWD = UserDefaults.standard.string(forKey: ChatSecure.Constants.keySaveTurnServerPWD) ?? ""
+		let turnServer = UserDefaults.standard.string(forKey: ChatSecure.Constants.keySaveTurnServer) ?? ""
+		let stunServer = UserDefaults.standard.string(forKey: ChatSecure.Constants.keySaveStunServer) ?? ""
 		// 8f87a00be37f0bfe19c0168ed0614966d70f2f8513ad66bda31a4a0a55fa89bd
 		// leZgnMWJMJ8QRFapC5liDHUrxJjalYqbhxPq+/V2zz8=
-		let stun = RTCIceServer(urlStrings: ["stun:stun.l.google.com:19302"])
-		let turn = RTCIceServer(urlStrings: ["turn:global.turn.twilio.com:3478"],
+		let stun = RTCIceServer(urlStrings: [stunServer])
+		let turn = RTCIceServer(urlStrings: [turnServer],
 								username: turnUser,
 								credential: turnPWD)
 		return [stun, turn]
 	}
 	
 	var peerConnection: RTCPeerConnection {
-		RTCInitializeSSL()
+//		RTCInitializeSSL()
 		if let peerConnection = _peerConnection {
 			return peerConnection
 		}
