+++ Packages/ChatSecure/ChatSecure/Classes/WebRTCJanus/JanusVideoRoom.swift
@@ -48,10 +48,10 @@ public class JanusVideoRoom: NSObject {
 	private var roomId: Int64 = 0
 	var useCustomCapturer = false
 	
-	init(delegate: JanusVideoRoomDelegate? = nil, token: String?) {
+	init(delegate: JanusVideoRoomDelegate? = nil, token: String?, rtcUrl: String) {
 		super.init()
 		//        let server = URL(string: "ws://54.235.68.160:8188/janus") // staging server
-		let server = URL(string: "ws://prod2.clearkeep.org:8188/janus") // dev server
+		let server = URL(string: rtcUrl) // dev server
 		//        let server = URL(string: "ws://10.0.255.82:8188/janus") // Local server
 		
 		//        let server = URL(string: AppConfig.buildEnvironment.webrtc)
